@@ -4,7 +4,7 @@ version: 2.0.0
 homepage: https://github.com/chinabrant/popup_menu
 
 environment:
-  sdk: ">=2.12.0 <3.0.0"
+  sdk: ">=2.17.5 <3.3.10"
 
 dependencies:
   flutter:
