@@ -1,4 +1,3 @@
-
 import 'package:flutter/material.dart';
 import 'package:popup_menu/popup_menu.dart';
 
@@ -7,7 +6,6 @@ class GestureDemo extends StatefulWidget {
   State<StatefulWidget> createState() {
     return _GestureDemoState();
   }
-
 }
 
 class _GestureDemoState extends State<GestureDemo> {
@@ -48,54 +46,51 @@ class _GestureDemoState extends State<GestureDemo> {
 
   void onShow() {
     PopupMenu menu = PopupMenu(
-        // backgroundColor: Colors.teal,
-        // lineColor: Colors.tealAccent,
-        // maxColumn: 2,
-        items: [
-          MenuItem(title: 'Copy', image: Image.asset('assets/copy.png')),
-          MenuItem(
-              title: 'Home',
-              // textStyle: TextStyle(fontSize: 10.0, color: Colors.tealAccent),
-              image: Icon(
-                Icons.home,
-                color: Colors.white,
-              )),
-          MenuItem(
-              title: 'Mail',
-              image: Icon(
-                Icons.mail,
-                color: Colors.white,
-              )),
-          MenuItem(
-              title: 'Power',
-              image: Icon(
-                Icons.power,
-                color: Colors.white,
-              )),
-          MenuItem(
-              title: 'Setting',
-              image: Icon(
-                Icons.settings,
-                color: Colors.white,
-              )),
-          MenuItem(
-              title: 'PopupMenu',
-              image: Icon(
-                Icons.menu,
-                color: Colors.white,
-              ))
-        ],
-        onClickMenu: onClickMenu,
-        // stateChanged: stateChanged,
-        onDismiss: onDismiss);
+      // backgroundColor: Colors.teal,
+      // lineColor: Colors.tealAccent,
+      // maxColumn: 2,
+      items: [
+        PopUpMenuItem(title: 'Copy', image: Image.asset('assets/copy.png')),
+        PopUpMenuItem(
+            title: 'Home',
+            // textStyle: TextStyle(fontSize: 10.0, color: Colors.tealAccent),
+            image: Icon(
+              Icons.home,
+              color: Colors.white,
+            )),
+        PopUpMenuItem(
+            title: 'Mail',
+            image: Icon(
+              Icons.mail,
+              color: Colors.white,
+            )),
+        PopUpMenuItem(
+            title: 'Power',
+            image: Icon(
+              Icons.power,
+              color: Colors.white,
+            )),
+        PopUpMenuItem(
+            title: 'Setting',
+            image: Icon(
+              Icons.settings,
+              color: Colors.white,
+            )),
+        PopUpMenuItem(
+            title: 'PopupMenu',
+            image: Icon(
+              Icons.menu,
+              color: Colors.white,
+            ))
+      ],
+      onClickMenu: onClickMenu,
+      // stateChanged: stateChanged,
+      onDismiss: onDismiss, context: context,
+    );
     menu.show(widgetKey: btnKey);
   }
 
-  void onClickMenu(MenuItemProvider item) {
+  void onClickMenu(MenuItemProvider item) {}
 
-  }
-
-  void onDismiss() {
-    
-  }
+  void onDismiss() {}
 }
